+++ src/main/java/com/bobi/timetracker/models/ProjectRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectRepository extends CrudRepository<Project, Integer> {
-    Project findProjectById(Integer id);
+public interface ProjectRepository extends CrudRepository<Project, Long> {
+    Project findProjectById(Long id);
 }
