+++ src/main/java/com/bobi/timetracker/models/RoleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends CrudRepository<Role, Integer> {
-    Role findRoleById(Integer id);
+public interface RoleRepository extends CrudRepository<Role, Long> {
+    Role findRoleById(Long id);
 }
