+++ src/main/java/com/bobi/timetracker/models/UserRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
     User findUserByUsername(String username);
-    User findUserById(Integer id);
+    User findUserById(Long id);
 }
